refactor memopedia builder into core module and shared config
